// frontend/src/app/core/api/model/cartItemDto.ts
/**
 * Gila E-Commerce API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProductDto } from './productDto';


export interface CartItemDto { 
    product: ProductDto;
    quantity: number;
}